gen_register_w: always cast the reset value to the correct type
It may happen that a register (e.g. the WDT CTRL register of Atmel ATSAMD21E17)
does not have the reset value specified.  In such cases the default
value 0u32 will be used.  However, if the register width is not 32 (as
is the case with the above mentioned WDT CTRL), rustc will complain.

So always cast the default value to the type that corresponds to the
register's width.