Handle missing dimIndex on registers
Per the SVD spec, dimIndex is not required:

> By default, <dimIndex> is a value starting with 0.
> Remark: Do not define <dimIndex> when using the placeholder [%s] in
<name> or <displayName>.

I interpret this as it being optional whether the name uses "%s" or
"[%s]". Atmel SAMD21 SVDs use "%s" with no dimIndex for 0-indexed
arrays.

This also tests that dimIndex is sequential and numeric in a single pass
without allocating temporary vectors, and fixes the typo
adresses -> addresses.