Revise union naming scheme
Rather than trying to compute a smart name and hoping that there is
no conflict, this commit switches to a simpler scheme: we now take
the shortest identifier from the components of a union and take the
one that sorts earliest as the name for the union in its containing
RegisterBlock.

So long as the names of the registers in that block don't have
collisions, we are guaranteed not to collide.