Remove #[deny(warnings)] from generated code
This declaration causes errors when using cargo doc, since some doc
comments include lines that give doc warnings. This doesn't prevent the
crate from building, so it's easy for these doc comment warnings to slip
through to releases. However, because of this line (which can't be
disabled later on), doc comment warnings are treated as errors and
prevent documentation for the generated crate from building, as well as
any other crate that depends on the generated crate.