remove dep on untagged unions
This removes any observable side effect of the --nightly switch by
removing the use of unions and the untagged_unions feature gate.
Unions are replaced with accessor functions that return the appropriate
register block reference.

Here's a playground link that shows that the pointer calculation looks
reasonable:

https://play.integer32.com/?version=stable&mode=debug&edition=2018&gist=cd56444abc03e6781e526bbddc7082bc

This commit is a breaking change.

This is based on this WIP PR branch:
https://github.com/rust-embedded/svd2rust/pull/256

It implements the easiest standalone portion of this issue:
https://github.com/rust-embedded/svd2rust/issues/218

and also makes accessing the unions "safe" too, which is requested here:
https://github.com/rust-embedded/svd2rust/issues/230