Basically revert #141 and mark methods #[inline(always)] again
Turns out this @therealprof dude was wrong in blindly following general
consensus that #[inline(always)] was a bad idea. Revisiting the topic
showed that just #[inline] is not enough to get even very trivial
functions inlined in dev mode which causes a ton of bloat and a lot of
debugging fun due to many emitted extra functions without any value for
the developer.

Usual binary reductions by this change are in the area of 10-15% which
is a lot given that even a simply blinky on Cortex-M0 is several kB
already.

E.g. before:

0.5% 100.0% 4.0KiB                .text section size, the file size is 734.1KiB

after:

0.5% 100.0% 3.7KiB                .text section size, the file size is 714.5KiB

Signed-off-by: Daniel Egger <daniel@eggers-club.de>