Add --mark_range flag
This flag allows implementing a marker trait for registers in specific
memory addresses. These can then be used to implement methods on
registers. One such example is providing atomic access to only some of
the registers directly in the PAC. Currently this is done in HAL, but
this loses the ability to use writer proxies. Moreover this allows
keeping atomic and non-atomic usage in HAL more similar.